Fase 2: Automated porting dengan 2to3
- jalankan 2to3 -w pada 45 file di src/ (source + test)
- fix TabError/IndentationError di sandbox/unicode_test.py
- fix long() -> int() di sandbox/common_gmvault.py
- verify: python -m compileall src lolos tanpa syntax error

+++ src/sandbox/common_gmvault.py
@@ -300,7 +300,7 @@ class IMAPBatchFetcher(object):
         if data == [None]: # no untagged responses...
             return [ ]
 
-        return [ long(i) for i in data[0].split() ]
+        return [ int(i) for i in data[0].split() ]
     
     def append(self, folder, msg, flags=(), msg_time=None):
         """Append a message to *folder*.
